requirements.txt
streamlit>=1.31.0
pymongo>=4.6.0
elasticsearch>=8.11.0
redis>=5.0.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
prefect>=2.14.0
kafka-python>=2.0.2
psycopg2-binary>=2.9.9
plotly>=5.18.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0